Update CI scripts

+++ ci/prepare.sh
@@ -7,7 +7,11 @@ set -ex
 WP_CLI_BIN_DIR=${WP_CLI_BIN_DIR-/tmp/wp-cli-phar}
 
 # Disable XDebug to speed up Composer and test suites.
-phpenv config-rm xdebug.ini
+if [ -f ~/.phpenv/versions/$(phpenv version-name)/etc/conf.d/xdebug.ini ]; then
+  phpenv config-rm xdebug.ini
+else
+  echo "xdebug.ini does not exist"
+fi
 
 composer install --no-interaction --prefer-source
 
@@ -20,7 +24,7 @@ then
 fi
 
 # the Behat test suite will pick up the executable found in $WP_CLI_BIN_DIR
-if [[ $BUILD == 'git' ]]
+if [[ $BUILD == 'git' || $BUILD == 'sniff' ]]
 then
 	echo $CLI_VERSION > VERSION
 else
